separated reading and converting gtfs feeds
