+++ src/main/java/org/matsim/pt2matsim/gtfs/GtfsConverter.java
@@ -16,114 +16,87 @@
  *                                                                         *
  * *********************************************************************** */
 
-
 package org.matsim.pt2matsim.gtfs;
 
-import com.opencsv.CSVReader;
 import org.apache.log4j.Logger;
 import org.matsim.api.core.v01.Coord;
 import org.matsim.api.core.v01.Id;
-import org.matsim.core.utils.collections.MapUtils;
 import org.matsim.core.utils.geometry.CoordinateTransformation;
 import org.matsim.core.utils.geometry.transformations.IdentityTransformation;
 import org.matsim.core.utils.geometry.transformations.TransformationFactory;
 import org.matsim.core.utils.misc.Time;
 import org.matsim.pt.transitSchedule.api.*;
-import org.matsim.pt2matsim.lib.RouteShape;
+import org.matsim.pt2matsim.gtfs.lib.*;
 import org.matsim.pt2matsim.lib.ShapedSchedule;
 import org.matsim.pt2matsim.lib.ShapedTransitSchedule;
-import org.matsim.pt2matsim.tools.*;
+import org.matsim.pt2matsim.tools.ScheduleTools;
 import org.matsim.vehicles.VehicleUtils;
 import org.matsim.vehicles.Vehicles;
-import org.matsim.pt2matsim.gtfs.lib.*;
 
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.IOException;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.time.LocalDate;
 import java.util.*;
-import java.util.Map.Entry;
-
 
 /**
- * Reads GTFS files and converts them to an unmapped MATSim Transit Schedule
- * <p/>
- * Based on GTFS2MATSimTransitSchedule by Sergio Ordonez
+ * Converts a GTFS feed to a MATSim transit schedule
  *
  * @author polettif
  */
-public class GtfsConverter implements GtfsFeed {
+public class GtfsConverter {
 
-	private static final Logger log = Logger.getLogger(GtfsConverter.class);
+	protected static Logger log = Logger.getLogger(GtfsConverter.class);
+
+	public static final String ALL_SERVICE_IDS = "all";
+	public static final String DAY_WITH_MOST_TRIPS = "dayWithMostTrips";
+	public static final String DAY_WITH_MOST_SERVICES = "dayWithMostServices";
+
+	private final GtfsFeed feed;
+
 
 	private boolean defaultAwaitDepartureTime = true;
 
 	private LocalDate dateUsed = null;
 
-	/**
-	 * Path to the folder where the gtfs files are located
-	 */
-	private String root;
-
-	/**
-	 * whether the gtfs feed uses frequencies.txt or not
-	 */
-	private boolean usesFrequencies = false;
-
-	/**
-	 * whether the gtfs feed uses shapes or not
-	 */
-	private boolean usesShapes = false;
+	private ShapedSchedule schedule;
+	private Vehicles vhcls;
 
 	/**
 	 * The time format used in the output MATSim transit schedule
 	 */
 	private SimpleDateFormat timeFormat = new SimpleDateFormat("HH:mm:ss");
 
-	/**
-	 * Set of service ids not defined in calendar.txt (only in calendar_dates.txt)
-	 */
-	private Set<String> serviceIdsNotInCalendarTxt = new HashSet<>();
 
-	/**
-	 * map for counting how many trips use each serviceId
-	 */
-	private Map<String, Integer> nTripsPerServiceId = new HashMap<>();
-
-
-
-	// containers for storing gtfs data
-	private Map<String, GtfsStop> gtfsStops = new HashMap<>();
-	private Map<String, GtfsRoute> gtfsRoutes = new TreeMap<>();
-	private Map<String, Service> services = new HashMap<>();
-	private Map<Id<RouteShape>, RouteShape> shapes = new HashMap<>();
-	private boolean warnStopTimes = true;
-
-	private final CoordinateTransformation transformation;
-	private ShapedTransitSchedule schedule = null;
-	private Vehicles vhcls = null;
-
-	public GtfsConverter(String gtfsFolder, String outputCoordinateSystem) {
-		this.transformation = TransformationFactory.getCoordinateTransformation("WGS84", outputCoordinateSystem);
-		loadFiles(gtfsFolder);
+	public GtfsConverter(GtfsFeed gtfsFeed) {
+		this.feed = gtfsFeed;
 	}
 
-	public GtfsConverter(String gtfsFolder) {
-		this.transformation = new IdentityTransformation();
-		loadFiles(gtfsFolder);
+
+	public void convert(String serviceIdsParam, String outputCoordinateSystem) {
+		convert(serviceIdsParam, TransformationFactory.getCoordinateTransformation("WGS84", outputCoordinateSystem), ScheduleTools.createSchedule(), VehicleUtils.createVehiclesContainer());
 	}
 
-	@Override
-	public void convert(String serviceIdsParam) {
-		convert(serviceIdsParam, ScheduleTools.createSchedule(), VehicleUtils.createVehiclesContainer());
+	public void convert(String serviceIdsParam, String outputCoordinateSystem, TransitSchedule transitSchedule, Vehicles vehicles) {
+		convert(serviceIdsParam, TransformationFactory.getCoordinateTransformation("WGS84", outputCoordinateSystem), transitSchedule, vehicles);
 	}
 
-	@Override
 	public void convert() {
-		convert(ALL_SERVICE_IDS, ScheduleTools.createSchedule(), VehicleUtils.createVehiclesContainer());
+		convert(ALL_SERVICE_IDS, new IdentityTransformation(), ScheduleTools.createSchedule(), VehicleUtils.createVehiclesContainer());
 	}
+
+	public TransitSchedule getSchedule() {
+		return schedule;
+	}
+
+	public Vehicles getVehicles() {
+		return vhcls;
+	}
+
+	public ShapedTransitSchedule getShapedTransitSchedule() {
+		return schedule;
+	}
+
+
 
 	/**
 	 * Converts the loaded gtfs data to a matsim transit schedule
@@ -136,8 +109,7 @@ public class GtfsConverter implements GtfsFeed {
 	 * <li>add transitRoute to the transitLine and thus to the schedule</li>
 	 * </ol>
 	 */
-	@Override
-	public void convert(String serviceIdsParam, TransitSchedule transitSchedule, Vehicles vehicles) {
+	public void convert(String serviceIdsParam, CoordinateTransformation transformation, TransitSchedule transitSchedule, Vehicles vehicles) {
 		log.info("#####################################");
 		log.info("Converting to MATSim transit schedule");
 
@@ -155,7 +127,6 @@ public class GtfsConverter implements GtfsFeed {
 
 		TransitScheduleFactory scheduleFactory = schedule.getFactory();
 
-
 		int counterLines = 0;
 		int counterRoutes = 0;
 
@@ -163,14 +134,14 @@ public class GtfsConverter implements GtfsFeed {
 		 * generating transitStopFacilities (mts) from gtfsStops and add them to the schedule.
 		 * Coordinates are transformed here.
 		 */
-		for(Entry<String, GtfsStop> stopEntry : gtfsStops.entrySet()) {
-			Coord result = transformation.transform(stopEntry.getValue().getPoint());
-			TransitStopFacility stopFacility = scheduleFactory.createTransitStopFacility(Id.create(stopEntry.getKey(), TransitStopFacility.class), result, stopEntry.getValue().isBlocks());
+		for(Map.Entry<String, Stop> stopEntry : feed.getStops().entrySet()) {
+			Coord stopFacilityCoord = transformation.transform(stopEntry.getValue().getPoint());
+			TransitStopFacility stopFacility = scheduleFactory.createTransitStopFacility(Id.create(stopEntry.getKey(), TransitStopFacility.class), stopFacilityCoord, stopEntry.getValue().isBlocks());
 			stopFacility.setName(stopEntry.getValue().getName());
 			schedule.addStopFacility(stopFacility);
 		}
 
-		if(usesFrequencies) {
+		if(feed.usesFrequencies()) {
 			log.info("    Using frequencies.txt to generate departures");
 		} else {
 			log.info("    Using stop_times.txt to generate departures");
@@ -178,7 +149,7 @@ public class GtfsConverter implements GtfsFeed {
 
 		DepartureIds departureIds = new DepartureIds();
 
-		for(GtfsRoute gtfsRoute : gtfsRoutes.values()) {
+		for(GtfsRoute gtfsRoute : feed.getRoutes().values()) {
 			/** [2]
 			 * Create a MTS transitLine for each GtfsRoute
 			 */
@@ -194,7 +165,7 @@ public class GtfsConverter implements GtfsFeed {
 
 				// if trip is part of used serviceId
 				for(String serviceId : serviceIdsToConvert) {
-					if(trip.getService().equals(services.get(serviceId))) {
+					if(trip.getService().equals(feed.getServices().get(serviceId))) {
 						isService = true;
 					}
 				}
@@ -236,7 +207,7 @@ public class GtfsConverter implements GtfsFeed {
 					 * Calculate departures from frequencies (if available)
 					 */
 					TransitRoute transitRoute = null;
-					if(usesFrequencies) {
+					if(feed.usesFrequencies()) {
 						transitRoute = scheduleFactory.createTransitRoute(Id.create(trip.getId(), TransitRoute.class), null, transitRouteStops, gtfsRoute.getRouteType().name);
 
 						for(Frequency frequency : trip.getFrequencies()) {
@@ -299,429 +270,6 @@ public class GtfsConverter implements GtfsFeed {
 	}
 
 	/**
-	 * Calls all methods to load the gtfs files. Order is critical
-	 */
-	private void loadFiles(String inputPath) {
-		if(!inputPath.endsWith("/")) inputPath += "/";
-		this.root = inputPath;
-
-		log.info("Loading GTFS files from " + root);
-		try { loadStops(); } catch (IOException e) {
-			throw new RuntimeException("File stops.txt not found!");
-		}
-		try { loadCalendar(); } catch (IOException e) {
-			throw new RuntimeException("File calendar.txt not found! ");
-		}
-		loadCalendarDates();
-		loadShapes();
-		try { loadRoutes(); } catch (IOException e) {
-			throw new RuntimeException("File routes.txt not found!");
-		}
-		try { loadTrips(); } catch (IOException e) {
-			throw new RuntimeException("File trips.txt not found!");
-		}
-		try { loadStopTimes(); } catch (IOException e) {
-			throw new RuntimeException("File stop_times.txt not found!");
-		}
-		loadFrequencies();
-		log.info("All files loaded");
-	}
-
-	/**
-	 * Reads all stops and puts them in {@link #gtfsStops}
-	 * <p/>
-	 * <br/><br/>
-	 * stops.txt <i>[https://developers.google.com/transit/gtfs/reference]</i><br/>
-	 * Individual locations where vehicles pick up or drop off passengers.
-	 *
-	 * @throws IOException
-	 */
-	private void loadStops() throws IOException {
-		log.info("Loading stops.txt");
-		CSVReader reader;
-
-		try {
-			reader = new CSVReader(new FileReader(root + GtfsDefinitions.Files.STOPS.fileName));
-			String[] header = reader.readNext(); // read header
-			Map<String, Integer> col = getIndices(header, GtfsDefinitions.Files.STOPS.columns, GtfsDefinitions.Files.STOPS.optionalColumns); // get column numbers for required fields
-
-			String[] line = reader.readNext();
-			while(line != null) {
-				Coord coord = new Coord(Double.parseDouble(line[col.get(GtfsDefinitions.STOP_LON)]), Double.parseDouble(line[col.get(GtfsDefinitions.STOP_LAT)]));
-				GtfsStop GtfsStop = new GtfsStop(coord, line[col.get(GtfsDefinitions.STOP_NAME)], false);
-				gtfsStops.put(line[col.get(GtfsDefinitions.STOP_ID)], GtfsStop);
-
-				line = reader.readNext();
-			}
-
-			reader.close();
-		} catch (ArrayIndexOutOfBoundsException i) {
-			throw new RuntimeException("Emtpy line found in stops.txt");
-		}
-		log.info("...     stops.txt loaded");
-	}
-
-	/**
-	 * Reads all services and puts them in {@link #services}
-	 * <p/>
-	 * <br/><br/>
-	 * calendar.txt <i>[https://developers.google.com/transit/gtfs/reference]</i><br/>
-	 * Dates for service IDs using a weekly schedule. Specify when service starts and ends,
-	 * as well as days of the week where service is available.
-	 *
-	 * @throws IOException
-	 */
-	private void loadCalendar() throws IOException {
-		log.info("Loading calendar.txt");
-		try {
-			CSVReader reader = new CSVReader(new FileReader(root + GtfsDefinitions.Files.CALENDAR.fileName));
-			String[] header = reader.readNext();
-			Map<String, Integer> col = getIndices(header, GtfsDefinitions.Files.CALENDAR.columns, GtfsDefinitions.Files.CALENDAR.optionalColumns);
-
-			// assuming all days really do follow monday in the file
-			int indexMonday = col.get("monday");
-
-			String[] line = reader.readNext();
-			int i = 1, c = 1;
-			while(line != null) {
-				if(i == Math.pow(2, c)) {
-					log.info("        # " + i);
-					c++;
-				}
-				i++;
-
-				boolean[] days = new boolean[7];
-				for(int d = 0; d < 7; d++) {
-					days[d] = line[indexMonday + d].equals("1");
-				}
-				services.put(line[col.get(GtfsDefinitions.SERVICE_ID)], new Service(line[col.get(GtfsDefinitions.SERVICE_ID)], days, line[col.get(GtfsDefinitions.START_DATE)], line[col.get(GtfsDefinitions.END_DATE)]));
-
-				line = reader.readNext();
-			}
-
-			reader.close();
-		} catch (ArrayIndexOutOfBoundsException i) {
-			throw new RuntimeException("Emtpy line found in calendar.txt");
-		}
-		log.info("...     calendar.txt loaded");
-	}
-
-	/**
-	 * Adds service exceptions to {@link #services} (if available)
-	 * <p/>
-	 * <br/><br/>
-	 * calendar_dates.txt <i>[https://developers.google.com/transit/gtfs/reference]</i><br/>
-	 * Exceptions for the service IDs defined in the calendar.txt file. If calendar_dates.txt includes ALL
-	 * dates of service, this file may be specified instead of calendar.txt.
-	 */
-	private void loadCalendarDates() {
-		// calendar dates are optional
-		log.info("Looking for calendar_dates.txt");
-		CSVReader reader;
-		try {
-			reader = new CSVReader(new FileReader(root + GtfsDefinitions.Files.CALENDAR_DATES.fileName));
-			String[] header = reader.readNext();
-			Map<String, Integer> col = getIndices(header, GtfsDefinitions.Files.CALENDAR_DATES.columns, GtfsDefinitions.Files.CALENDAR_DATES.optionalColumns);
-
-			String[] line = reader.readNext();
-			while(line != null) {
-				Service currentService = services.get(line[col.get(GtfsDefinitions.SERVICE_ID)]);
-
-				if(currentService == null) {
-					currentService = new Service(
-							line[col.get(GtfsDefinitions.SERVICE_ID)],
-							new boolean[] { false, false, false, false, false, false, false },
-							"19700101", "29991231"
-					);
-
-					services.put(currentService.getId(), currentService);
-
-					if(serviceIdsNotInCalendarTxt.add(currentService.getId())) {
-						log.warn("Service id \"" + currentService.getId() + "\" not defined in calendar.txt, only in calendar_dates.txt. Service id will still be used.");
-					}
-				}
-
-				if(line[col.get(GtfsDefinitions.EXCEPTION_TYPE)].equals("2")) {
-					currentService.addException(line[col.get(GtfsDefinitions.DATE)]);
-				} else {
-					currentService.addAddition(line[col.get(GtfsDefinitions.DATE)]);
-				}
-
-				line = reader.readNext();
-			}
-			reader.close();
-			log.info("...     calendar_dates.txt loaded");
-		} catch (IOException e) {
-			log.info("...     no calendar dates file found.");
-		} catch (ArrayIndexOutOfBoundsException i) {
-			throw new RuntimeException("Emtpy line found in calendar_dates.txt");
-		}
-	}
-
-	/**
-	 * Loads shapes (if available) and puts them in {@link #shapes}. A shape is a sequence of points, i.e. a line.
-	 * <p/>
-	 * <br/><br/>
-	 * shapes.txt <i>[https://developers.google.com/transit/gtfs/reference]</i><br/>
-	 * Rules for drawing lines on a map to represent a transit organization's routes.
-	 */
-	private void loadShapes() {
-		// shapes are optional
-		log.info("Looking for shapes.txt");
-		CSVReader reader;
-		try {
-			reader = new CSVReader(new FileReader(root + GtfsDefinitions.Files.SHAPES.fileName));
-
-			String[] header = reader.readNext();
-			Map<String, Integer> col = getIndices(header, GtfsDefinitions.Files.SHAPES.columns, GtfsDefinitions.Files.SHAPES.optionalColumns);
-
-			String[] line = reader.readNext();
-			while(line != null) {
-				usesShapes = true; // shape file might exists but could be empty
-
-				Id<RouteShape> shapeId = Id.create(line[col.get(GtfsDefinitions.SHAPE_ID)], RouteShape.class);
-
-				RouteShape currentShape = shapes.get(shapeId);
-				if(currentShape == null) {
-					currentShape = new GtfsShape(line[col.get(GtfsDefinitions.SHAPE_ID)]);
-					shapes.put(shapeId, currentShape);
-				}
-				Coord point = new Coord(Double.parseDouble(line[col.get(GtfsDefinitions.SHAPE_PT_LON)]), Double.parseDouble(line[col.get(GtfsDefinitions.SHAPE_PT_LAT)]));
-				currentShape.addPoint(transformation.transform(point), Integer.parseInt(line[col.get(GtfsDefinitions.SHAPE_PT_SEQUENCE)]));
-				line = reader.readNext();
-			}
-			reader.close();
-			log.info("...     shapes.txt loaded");
-		} catch (IOException e) {
-			log.info("...     no shapes file found.");
-		} catch (ArrayIndexOutOfBoundsException i) {
-			throw new RuntimeException("Emtpy line found in shapes.txt");
-		}
-	}
-
-	/**
-	 * Basically just reads all routeIds and their corresponding names and types and puts them in {@link #gtfsRoutes}.
-	 * <p/>
-	 * <br/><br/>
-	 * routes.txt <i>[https://developers.google.com/transit/gtfs/reference]</i><br/>
-	 * Transit routes. A route is a group of trips that are displayed to riders as a single service.
-	 *
-	 * @throws IOException
-	 */
-	private void loadRoutes() throws IOException {
-		log.info("Loading routes.txt");
-		try {
-			CSVReader reader = new CSVReader(new FileReader(root + GtfsDefinitions.Files.ROUTES.fileName));
-			String[] header = reader.readNext();
-			Map<String, Integer> col = getIndices(header, GtfsDefinitions.Files.ROUTES.columns, GtfsDefinitions.Files.ROUTES.optionalColumns);
-
-			String[] line = reader.readNext();
-			while(line != null) {
-				int routeTypeNr = Integer.parseInt(line[col.get(GtfsDefinitions.ROUTE_TYPE)]);
-				if(routeTypeNr < 0 || routeTypeNr > 7) {
-					throw new RuntimeException("Invalid value for route type: " + routeTypeNr + " [https://developers.google.com/transit/gtfs/reference/routes-file]");
-				}
-
-				GtfsRoute newGtfsRoute = new GtfsRoute(line[col.get(GtfsDefinitions.ROUTE_ID)], line[col.get(GtfsDefinitions.ROUTE_SHORT_NAME)], GtfsDefinitions.RouteTypes.values()[routeTypeNr]);
-				gtfsRoutes.put(line[col.get(GtfsDefinitions.ROUTE_ID)], newGtfsRoute);
-
-				line = reader.readNext();
-			}
-			reader.close();
-		} catch (ArrayIndexOutOfBoundsException i) {
-			throw new RuntimeException("Emtpy line found in routes.txt");
-		}
-		log.info("...     routes.txt loaded");
-	}
-
-	/**
-	 * Generates a trip with trip_id and adds it to the corresponding route (referenced by route_id) in {@link #gtfsRoutes}.
-	 * Adds the shape_id as well (if shapes are used). Each trip uses one service_id, the serviceIds statistics are increased accordingly
-	 * <p/>
-	 * <br/><br/>
-	 * trips.txt <i>[https://developers.google.com/transit/gtfs/reference]</i><br/>
-	 * Trips for each route. A trip is a sequence of two or more gtfsStops that occurs at specific time.
-	 *
-	 * @throws IOException
-	 */
-	private void loadTrips() throws IOException {
-		log.info("Loading trips.txt");
-		try {
-			CSVReader reader = new CSVReader(new FileReader(root + GtfsDefinitions.Files.TRIPS.fileName));
-			String[] header = reader.readNext();
-			Map<String, Integer> col = getIndices(header, GtfsDefinitions.Files.TRIPS.columns, GtfsDefinitions.Files.TRIPS.optionalColumns);
-
-			String[] line = reader.readNext();
-			while(line != null) {
-				GtfsRoute gtfsRoute = gtfsRoutes.get(line[col.get("route_id")]);
-				if(usesShapes) {
-					Id<RouteShape> shapeId = Id.create(line[col.get(GtfsDefinitions.SHAPE_ID)], RouteShape.class);
-					Trip newTrip = new Trip(line[col.get(GtfsDefinitions.TRIP_ID)], services.get(line[col.get(GtfsDefinitions.SERVICE_ID)]), shapes.get(shapeId), line[col.get(GtfsDefinitions.TRIP_ID)]);
-					gtfsRoute.putTrip(line[col.get(GtfsDefinitions.TRIP_ID)], newTrip);
-				} else {
-					Trip newTrip = new Trip(line[col.get(GtfsDefinitions.TRIP_ID)], services.get(line[col.get(GtfsDefinitions.SERVICE_ID)]), null, line[col.get(GtfsDefinitions.TRIP_ID)]);
-					gtfsRoute.putTrip(line[col.get(GtfsDefinitions.TRIP_ID)], newTrip);
-				}
-
-				// each trip uses one service id, increase statistics accordingly
-				Integer count = MapUtils.getInteger(line[col.get(GtfsDefinitions.SERVICE_ID)], nTripsPerServiceId, 1);
-				nTripsPerServiceId.put(line[col.get(GtfsDefinitions.SERVICE_ID)], count + 1);
-
-				line = reader.readNext();
-			}
-
-			reader.close();
-		} catch (ArrayIndexOutOfBoundsException i) {
-			throw new RuntimeException("Emtpy line found in trips.txt");
-		}
-		log.info("...     trips.txt loaded");
-	}
-
-	/**
-	 * Stop times are added to their respective trip (which are stored in {@link #gtfsRoutes}).
-	 * <p/>
-	 * <br/><br/>
-	 * stop_times.txt <i>[https://developers.google.com/transit/gtfs/reference]</i><br/>
-	 * Times that a vehicle arrives at and departs from individual gtfsStops for each trip.
-	 *
-	 * @throws IOException
-	 */
-	private void loadStopTimes() throws IOException {
-		log.info("Loading stop_times.txt");
-		try {
-			CSVReader reader = new CSVReader(new FileReader(root + GtfsDefinitions.Files.STOP_TIMES.fileName));
-			String[] header = reader.readNext();
-			Map<String, Integer> col = getIndices(header, GtfsDefinitions.Files.STOP_TIMES.columns, GtfsDefinitions.Files.STOP_TIMES.optionalColumns);
-
-			String[] line = reader.readNext();
-			int i = 1, c = 1;
-			while(line != null) {
-				if(i == Math.pow(2, c)) { log.info("        # " + i); c++; } i++; // just for logging so something happens in the console
-
-				for(GtfsRoute currentGtfsRoute : gtfsRoutes.values()) {
-					Trip trip = currentGtfsRoute.getTrips().get(line[col.get(GtfsDefinitions.TRIP_ID)]);
-					if(trip != null) {
-						try {
-							if(!line[col.get(GtfsDefinitions.ARRIVAL_TIME)].equals("")) {
-								trip.putStopTime(
-									Integer.parseInt(line[col.get(GtfsDefinitions.STOP_SEQUENCE)]),
-									new StopTime(Integer.parseInt(line[col.get(GtfsDefinitions.STOP_SEQUENCE)]),
-											timeFormat.parse(line[col.get(GtfsDefinitions.ARRIVAL_TIME)]),
-											timeFormat.parse(line[col.get(GtfsDefinitions.DEPARTURE_TIME)]),
-											line[col.get(GtfsDefinitions.STOP_ID)]));
-							}
-							/** GTFS Reference: If this stop isn't a time point, use an empty string value for the
-							 * arrival_time and departure_time fields.
-							 */
-							else {
-								trip.putStop(Integer.parseInt(line[col.get(GtfsDefinitions.STOP_SEQUENCE)]), line[col.get(GtfsDefinitions.STOP_ID)]);
-								if(warnStopTimes) {
-									log.warn("No arrival time set! Stops without arrival times will be scheduled based on the " +
-											"nearest preceding timed stop. This message is only given once.");
-									warnStopTimes = false;
-								}
-							}
-						} catch (NumberFormatException | ParseException e) {
-							e.printStackTrace();
-						}
-					}
-				}
-				line = reader.readNext();
-			}
-
-			reader.close();
-		} catch (ArrayIndexOutOfBoundsException i) {
-			throw new RuntimeException("Emtpy line found in stop_times.txt");
-		}
-		log.info("...     stop_times.txt loaded");
-	}
-
-	/**
-	 * Loads the frequencies (if available) and adds them to their respective trips in {@link #gtfsRoutes}.
-	 * <p/>
-	 * <br/><br/>
-	 * frequencies.txt <i>[https://developers.google.com/transit/gtfs/reference]</i><br/>
-	 * Headway (time between trips) for gtfsRoutes with variable frequency of service.
-	 */
-	private void loadFrequencies() {
-		log.info("Looking for frequencies.txt");
-		// frequencies are optional
-		CSVReader reader;
-		try {
-			reader = new CSVReader(new FileReader(root + GtfsDefinitions.Files.FREQUENCIES.fileName));
-			String[] header = reader.readNext();
-			Map<String, Integer> col = getIndices(header, GtfsDefinitions.Files.FREQUENCIES.columns, GtfsDefinitions.Files.FREQUENCIES.optionalColumns);
-
-			String[] line = reader.readNext();
-			while(line != null) {
-				usesFrequencies = true;    // frequencies file might exists but could be empty
-
-				for(GtfsRoute actualGtfsRoute : gtfsRoutes.values()) {
-					Trip trip = actualGtfsRoute.getTrips().get(line[col.get(GtfsDefinitions.TRIP_ID)]);
-					if(trip != null) {
-						try {
-							trip.addFrequency(new Frequency(timeFormat.parse(line[col.get(GtfsDefinitions.START_TIME)]), timeFormat.parse(line[col.get(GtfsDefinitions.END_TIME)]), Integer.parseInt(line[col.get(GtfsDefinitions.HEADWAY_SECS)])));
-						} catch (NumberFormatException | ParseException e) {
-							e.printStackTrace();
-						}
-					}
-				}
-				line = reader.readNext();
-			}
-			reader.close();
-			log.info("...     frequencies.txt loaded");
-		} catch (FileNotFoundException e) {
-			log.info("...     no frequencies file found.");
-		} catch (IOException e) {
-			e.printStackTrace();
-		} catch (ArrayIndexOutOfBoundsException e) {
-			throw new RuntimeException("Emtpy line found in frequencies.txt");
-		}
-	}
-
-	/**
-	 * In case optional columns in a csv file are missing or are out of order, addressing array
-	 * values directly via integer (i.e. where the column should be) does not work.
-	 *
-	 * @param header      the header (first line) of the csv file
-	 * @param requiredColumns array of attributes you need the indices of
-	 * @return the index for each attribute given in columnNames
-	 */
-	private static Map<String, Integer> getIndices(String[] header, String[] requiredColumns, String[] optionalColumns) {
-		Map<String, Integer> indices = new HashMap<>();
-		Set<String> requiredNotFound = new HashSet<>();
-
-		for(String columnName : requiredColumns) {
-			boolean found = false;
-			for(int i = 0; i < header.length; i++) {
-				if(header[i].equals(columnName)) {
-					indices.put(columnName, i);
-					found = true;
-					break;
-				}
-			}
-			if(!found) {
-				requiredNotFound.add(columnName);
-			}
-		}
-
-		if(requiredNotFound.size() > 0) {
-			throw new IllegalArgumentException("Required column(s) "+requiredNotFound+" not found in csv. Might be some additional characters in the header or the encoding not being UTF-8.");
-		}
-
-		for(String columnName : optionalColumns) {
-			for(int i = 0; i < header.length; i++) {
-				if(header[i].equals(columnName)) {
-					indices.put(columnName, i);
-					break;
-				}
-			}
-		}
-		return indices;
-	}
-
-
-	/**
 	 * @return The date from which services and thus trips should be extracted
 	 */
 	private LocalDate getExtractDate(String param) {
@@ -736,7 +284,7 @@ public class GtfsConverter implements GtfsFeed {
 				log.info("    Using service IDs of the day with the most services (" + DAY_WITH_MOST_SERVICES + ").");
 				LocalDate date = null;
 				int maxNServiceIds = 0;
-				for(Entry<LocalDate, Set<String>> idsOnDayEntry : Service.dateStats.entrySet()) {
+				for(Map.Entry<LocalDate, Set<String>> idsOnDayEntry : Service.dateStats.entrySet()) {
 					try {
 						LocalDate currentDate = idsOnDayEntry.getKey();
 						Set<String> currentIds = idsOnDayEntry.getValue();
@@ -755,10 +303,10 @@ public class GtfsConverter implements GtfsFeed {
 				log.info("    Using service IDs of the day with the most trips (" + DAY_WITH_MOST_TRIPS + ").");
 				LocalDate date = null;
 				int maxTrips = 0;
-				for(Entry<LocalDate, Set<String>> e : Service.dateStats.entrySet()) {
+				for(Map.Entry<LocalDate, Set<String>> e : Service.dateStats.entrySet()) {
 					int nTrips = 0;
 					for(String s : e.getValue()) {
-						nTrips += nTripsPerServiceId.get(s);
+						nTrips += feed.getNTripsPerServiceId().get(s);
 					}
 					if(nTrips > maxTrips) {
 						maxTrips = nTrips;
@@ -782,10 +330,10 @@ public class GtfsConverter implements GtfsFeed {
 
 	private Set<String> getServiceIds(LocalDate checkDate) {
 		if(checkDate == null) {
-			return services.keySet();
+			return new HashSet<>(feed.getServices().keySet());
 		} else {
 			HashSet<String> idsOnCheckDate = new HashSet<>();
-			for(Service service : services.values()) {
+			for(Service service : feed.getServices().values()) {
 				if(serviceCoversDate(service, checkDate)) {
 					idsOnCheckDate.add(service.getId());
 				}
@@ -814,16 +362,6 @@ public class GtfsConverter implements GtfsFeed {
 		return false;
 	}
 
-	@Override
-	public Map<Id<RouteShape>, RouteShape> getShapes() {
-		return shapes;
-	}
-
-
-	public Map<String, GtfsRoute> getGtfsRoutes() {
-		return gtfsRoutes;
-	}
-
 	/**
 	 * helper class for meaningful departureIds
 	 */
@@ -842,18 +380,4 @@ public class GtfsConverter implements GtfsFeed {
 
 		}
 	}
-
-	public ShapedTransitSchedule getShapedTransitSchedule() {
-		return schedule;
 	}
-
-	public TransitSchedule getSchedule() {
-		return schedule;
-	}
-
-	public Vehicles getVehicles() {
-		return vhcls;
-	}
-
-
-}
